moving Importer to components/importer.py
