+++ src/datumaro/plugins/data_formats/ade20k2020.py
@@ -17,7 +17,8 @@ from datumaro.components.annotation import (
     Mask,
     Polygon,
 )
-from datumaro.components.dataset_base import DatasetBase, DatasetItem, Importer
+from datumaro.components.dataset_base import DatasetBase, DatasetItem
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
 from datumaro.util import parse_json
