+++ src/datumaro/plugins/data_formats/brats_numpy.py
@@ -7,7 +7,8 @@ import os.path as osp
 import numpy as np
 
 from datumaro.components.annotation import AnnotationType, Cuboid3d, LabelCategories, Mask
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import MultiframeImage
 from datumaro.util.pickle_util import PickleLoader
