+++ src/datumaro/plugins/data_formats/datumaro/base.py
@@ -20,7 +20,8 @@ from datumaro.components.annotation import (
     RleMask,
     Skeleton,
 )
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.errors import DatasetImportError, InvalidAnnotationError
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image, MediaElement, PointCloud
