+++ src/datumaro/plugins/data_formats/image_dir.py
@@ -6,7 +6,8 @@ import logging as log
 import os
 import os.path as osp
 
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.exporter import Exporter
 from datumaro.components.media import Image
 from datumaro.util.image import find_images
