+++ src/datumaro/plugins/data_formats/image_zip.py
@@ -8,7 +8,8 @@ import os.path as osp
 from enum import Enum
 from zipfile import ZIP_BZIP2, ZIP_DEFLATED, ZIP_LZMA, ZIP_STORED, ZipFile
 
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.exporter import Exporter
 from datumaro.components.media import ByteImage
 from datumaro.util import parse_str_enum_value
