+++ src/datumaro/plugins/data_formats/kitti_raw/base.py
@@ -8,7 +8,8 @@ import os.path as osp
 from defusedxml import ElementTree as ET
 
 from datumaro.components.annotation import AnnotationType, Cuboid3d, LabelCategories
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image, PointCloud
 from datumaro.util import cast
