+++ src/datumaro/plugins/data_formats/lfw.py
@@ -7,7 +7,8 @@ import os.path as osp
 import re
 
 from datumaro.components.annotation import AnnotationType, Label, LabelCategories, Points
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.errors import MediaTypeError
 from datumaro.components.exporter import Exporter
 from datumaro.components.format_detection import FormatDetectionContext
