+++ src/datumaro/plugins/data_formats/market1501.py
@@ -6,7 +6,8 @@ import os
 import os.path as osp
 import re
 
-from datumaro.components.dataset_base import DatasetBase, DatasetItem, Importer
+from datumaro.components.dataset_base import DatasetBase, DatasetItem
+from datumaro.components.importer import Importer
 from datumaro.components.errors import MediaTypeError
 from datumaro.components.exporter import Exporter
 from datumaro.components.media import Image
