+++ src/datumaro/plugins/data_formats/mots.py
@@ -13,7 +13,8 @@ from glob import iglob
 import numpy as np
 
 from datumaro.components.annotation import AnnotationType, LabelCategories, Mask
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.errors import MediaTypeError
 from datumaro.components.exporter import Exporter
 from datumaro.components.media import Image
