+++ src/datumaro/plugins/data_formats/mpii/mpii_json.py
@@ -7,7 +7,8 @@ import os.path as osp
 import numpy as np
 
 from datumaro.components.annotation import Bbox, LabelCategories, Points, PointsCategories
-from datumaro.components.dataset_base import AnnotationType, DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import AnnotationType, DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
 from datumaro.util import parse_json_file
