+++ src/datumaro/plugins/data_formats/nyu_depth_v2.py
@@ -9,7 +9,8 @@ import h5py
 import numpy as np
 
 from datumaro.components.annotation import DepthAnnotation
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
