+++ src/datumaro/plugins/data_formats/sly_pointcloud/base.py
@@ -6,7 +6,8 @@ import os.path as osp
 from glob import iglob
 
 from datumaro.components.annotation import AnnotationType, Cuboid3d, LabelCategories
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.media import Image, PointCloud
 from datumaro.util import parse_json_file
 from datumaro.util.image import find_images
