+++ src/datumaro/plugins/data_formats/video.py
@@ -5,7 +5,8 @@
 import os.path as osp
 from typing import Optional
 
-from datumaro.components.dataset_base import DatasetBase, DatasetItem, Importer
+from datumaro.components.dataset_base import DatasetBase, DatasetItem
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionConfidence, FormatDetectionContext
 from datumaro.components.media import Video, VideoFrame
 from datumaro.util.definitions import DEFAULT_SUBSET_NAME
