+++ src/datumaro/plugins/data_formats/voc/importer.py
@@ -4,7 +4,7 @@
 
 import os.path as osp
 
-from datumaro.components.dataset_base import Importer
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 
 from .format import VocPath, VocTask
