+++ src/datumaro/plugins/data_formats/vott_csv.py
@@ -6,7 +6,8 @@ import csv
 import os.path as osp
 
 from datumaro.components.annotation import AnnotationType, Bbox, LabelCategories
-from datumaro.components.dataset_base import DatasetItem, Importer, SubsetBase
+from datumaro.components.dataset_base import DatasetItem, SubsetBase
+from datumaro.components.importer import Importer
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
 from datumaro.util.meta_file_util import has_meta_file, parse_meta_file
